backend-django/requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
PyMySQL==1.1.2
requests==2.32.5
python-dotenv==1.0.0
cryptography==43.0.3
gunicorn==21.2.0